pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "zdl_nlp"
description = "Natural Language Processing Pipeline @ DWDS/ZDL"
authors = [
    {name = "Gregor Middell", email = "gregor.middell@bbaw.de"},
    {name = "Luise Köhler", email = "luise.koehler@bbaw.de"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Topic :: Education",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing :: Linguistic"
]
requires-python = ">=3.12"
dynamic = ["readme", "version"]
dependencies = [
    "backoff",
    "conllu",
    "datasketch",
    "de_zdl_lg",
    "dwdsmor-dwds",
    "gdex @ git+https://github.com/zentrum-lexikographie/gdex.git@v1.7.0",
    "lingua-language-detector",
    "lxml",
    "pgvector",
    "pika",
    "psycopg[binary]",
    "python-dotenv",
    "python-dateutil",
    "ratelimit",
    "requests",
    "retry",
    "SoMaJo",
    "spacy",
    "tqdm"
]

[project.optional-dependencies]
gpu = [
    "de_zdl_dist",
    "dwds-wic-sbert",
    "spacy[cuda12x,transformers]",
]
dev = [
    "flake8",
    "flake8-bugbear",
    "isort",
    "black",
    "mypy",
    "pre-commit",
    "pytest"
]

[project.urls]
Homepage = "https://github.com/zentrum-lexikographie/nlp-pipeline"

[project.scripts]
zdl-nlp-annotate = "zdl_nlp.annotate:main"
zdl-nlp-colloc = "zdl_nlp.colloc:main"
zdl-nlp-ddc = "zdl_nlp.ddc.corpora:main"
zdl-nlp-ddc2conllu = "zdl_nlp.ddc.tabs:main"
zdl-nlp-examples = "zdl_nlp.examples:main"
zdl-nlp-install-models = "zdl_nlp.models:install"
zdl-nlp-korap = "zdl_nlp.korap:main"
zdl-nlp-polspeech = "zdl_nlp.polspeech:main"
zdl-nlp-sbert = "zdl_nlp.sbert:main"
zdl-nlp-txt2conllu = "zdl_nlp.segment:main"
zdl-nlp-tei2conllu = "zdl_nlp.tei:main"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::FutureWarning",
]

[tool.mypy]
disable_error_code = "import-untyped"
ignore_missing_imports = true

[tool.setuptools.dynamic]
readme = {file = ["README.md"], content-type = "text/markdown"}
version = {attr = "zdl_nlp.version.__version__"}

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests"]
